+++ test/Unit/Exception/ResolvedToRootSchemaTest.php
@@ -28,7 +28,7 @@ final class ResolvedToRootSchemaTest extends Framework\TestCase
 {
     public function testJsonPointerReturnsException(): void
     {
-        $jsonPointer = JsonPointer::fromString('/foo/bar');
+        $jsonPointer = JsonPointer::fromString('#/foo/bar');
 
         $exception = Exception\ResolvedToRootSchema::jsonPointer($jsonPointer);