+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`2.0.0...main`][2.0.0...main].
 ### Changed
 
 - Required [`ergebnis/json-pointer`](https://github.com/ergebnis/json-pointer) ([#195]), by [@localheinz]
+- Started throwing an `Exception\CanNotResolve` exception instead of an `Exception\ResolvedToRootSchema` when the `JsonPointer` is not a valid URI fragment identifier representation of a JSON pointer ([#202]), by [@localheinz]
 
 ## [`2.0.0`][2.0.0]
 
@@ -71,5 +72,6 @@ For a full diff see [`dcd4cfb...1.0.0`][dcd4cfb...1.0.0].
 [#169]: https://github.com/ergebnis/json-schema-validator/pull/169
 [#172]: https://github.com/ergebnis/json-schema-validator/pull/172
 [#195]: https://github.com/ergebnis/json-schema-validator/pull/195
+[#202]: https://github.com/ergebnis/json-schema-validator/pull/202
 
 [@localheinz]: https://github.com/localheinz